Rename seqimporter -> ocs2
