+++ modules/ocs2/src/main/scala/gem/ocs2/Importer.scala
@@ -1,4 +1,4 @@
-package gem.seqimporter
+package gem.ocs2
 
 import doobie.imports._
