+++ modules/ocs2/src/main/scala/gem/ocs2/package.scala
@@ -2,7 +2,7 @@ package gem
 
 import scalaz._
 
-package object seqimporter {
+package object ocs2 {
   type ConfigMap = String ==>> String
 
   val EmptyConfigMap = ==>>.empty[String, String]
