+++ modules/ocs2/src/main/scala/gem/ocs2/pio/package.scala
@@ -1,4 +1,4 @@
-package gem.seqimporter
+package gem.ocs2
 
 package object pio {
   type PioParse[A] = String => Option[A]
